+++ contracts/PublicSha256Spt.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.7.0;
 import "./PublicSpt.sol";
 
 contract PublicSha256Spt is PublicSpt {
-    constructor(uint _depth) PublicSpt(_depth) public {}
+    constructor(uint8 _depth) PublicSpt(_depth) public {}
 
     function hash(bytes memory data) override internal returns (bytes32) {
         return sha256(data);
